requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.9.1
python-dotenv==0.21.0
pytz==2022.5
sqlparse==0.4.3
toml==0.10.2
